refactored range_gather with ut updated
