+++ magi_attention/common/range_op/_range_gather.py
@@ -19,7 +19,27 @@ import triton.language as tl
 
 from magi_attention.utils import nvtx
 
+from ..ranges import NaiveRanges
+
 __all__ = ["range_gather"]
+
+
+def _calc_cu_range_sizes(
+    ranges: torch.Tensor | NaiveRanges,
+    device: int,
+) -> tuple[torch.Tensor, int]:
+    if isinstance(ranges, torch.Tensor):
+        ranges = ranges.tolist()
+
+    cu_range_sizes = []
+    total_size = 0
+    for start, end in ranges:
+        cu_range_sizes.append(total_size)
+        total_size += end - start
+
+    cu_range_sizes = torch.tensor(cu_range_sizes, dtype=torch.int32, device=device)
+
+    return cu_range_sizes, total_size
 
 
 def _calc_ranges_row_map(
@@ -83,27 +103,52 @@ def range_gather_kernel(
 def range_gather(
     input: torch.Tensor,
     ranges: torch.Tensor,
-    cu_range_sizes: torch.Tensor,
-    total_size: int,
     dim: int = 0,
-    row_map: torch.Tensor | None = None,
     output: torch.Tensor | None = None,
+    **kwargs,
 ) -> torch.Tensor:
     """
     Gather values from input tensor based on specified ranges into a new output tensor.
 
     Args:
-        input: Source tensor to gather from
-        ranges: Tensor of [start, end] ranges in the input
-        cu_range_sizes: Cumulative sizes of ranges
-        total_size: Total number of rows in the output tensor
+        input (torch.Tensor): Source tensor to gather from
+        ranges (torch.Tensor): Tensor of [start, end] ranges in the input
         dim: Dimension along which to perform the gather operation
-        row_map: Optional mapping from row indices to range indices
         output: Optional output tensor buffer to store the result
+
+        kwargs:
+            - cu_range_sizes: Cumulative sizes of ranges
+            - total_size: Total number of rows in the output tensor
+            - row_map: Optional mapping from row indices to range indices
 
     Returns:
         A new tensor containing the gathered values, put into output if provided.
     """
+
+    # ---   calculate meta   --- #
+
+    # Make ranges contiguous
+    ranges = ranges.contiguous()
+
+    # Calculate cu_range_sizes and total_size if not provided
+    cu_range_sizes = kwargs.pop("cu_range_sizes", None)
+    total_size = kwargs.pop("total_size", None)
+    if cu_range_sizes is None or total_size is None:
+        cu_range_sizes, total_size = _calc_cu_range_sizes(
+            ranges,
+            device=input.device,
+        )
+    else:
+        cu_range_sizes = cu_range_sizes.contiguous()
+
+    # Calculate row_map if not provided
+    row_map = kwargs.pop("row_map", None)
+    if row_map is None:
+        row_map = _calc_ranges_row_map(ranges, total_size)
+    else:
+        row_map = row_map.contiguous()
+
+    # ---   pre-process input/output   --- #
 
     if output is None:
         output_shape = list(input.shape)
@@ -125,19 +170,12 @@ def range_gather(
         input = input.contiguous()
         output = output.contiguous()
 
-    ranges = ranges.contiguous()
-    cu_range_sizes = cu_range_sizes.contiguous()
-    # Calculate row_map if not provided
-    if row_map is None:
-        row_map = _calc_ranges_row_map(ranges, total_size)
-    else:
-        row_map = row_map.contiguous()
-
     # Calculate stride (considering memory step size of elements)
     input_stride = input.stride(0)
     output_stride = output.stride(0)
 
-    # Calculate grid size
+    # ---   calculate grid size   --- #
+
     M = total_size
     N = input.numel() // input.shape[0]
 
@@ -146,7 +184,8 @@ def range_gather(
 
     grid = (M, N_BLOCK)
 
-    # Launch kernel
+    # ---   launch kernel   --- #
+
     range_gather_kernel[grid](
         input,
         output,
@@ -159,6 +198,8 @@ def range_gather(
         N_BLOCK,
         ELEM_PER_BLOCK,
     )
+
+    # ---   post-process output   --- #
 
     # If transposed earlier, transpose back
     if dim != 0:
